extract stories
