+++ packages/react-positioning/src/Positioning.stories.tsx
@@ -1,0 +1,22 @@
+import { Positioning } from './utils.stories';
+import descriptionMd from './PositioningDescription.md';
+import bestPracticesMd from './PositioningBestPractices.md';
+
+export { Default } from './PositioningDefault.stories';
+export { ShorthandPositions } from './PositioningShorthandPositions.stories';
+export { CoverTarget } from './PositioningCoverTarget.stories';
+export { Offset } from './PositioningOffset.stories';
+export { AnchorToTarget } from './PositioningAnchorToTarget.stories';
+
+export default {
+  title: 'Concepts/Positioning',
+  component: Positioning,
+  parameters: {
+    layout: 'centered',
+    docs: {
+      description: {
+        component: [descriptionMd, bestPracticesMd].join('\n'),
+      },
+    },
+  },
+};
