+++ packages/react-positioning/src/PositioningBestPractices.md
@@ -1,0 +1,6 @@
+# Best practices
+
+⚠️ Simpler configuration is better. Positioning is complicated, do not make your life harder by using it too much ⚠️
+
+These examples are intended to document the `positioning` prop used in Fluent, please refer to component specific
+documentation for best practices for a specific component.
