+++ packages/react-positioning/src/PositioningDefault.stories.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+import { PositioningProps } from './types';
+import { PositionedComponent } from './utils.stories';
+
+export const Default = (props: PositioningProps) => {
+  return <PositionedComponent positioning={props} />;
+};
