--- packages/react-positioning/src/positioningBestPractices.md
@@ -1,3 +1,0 @@
-# Best practices
-
-⚠️ Simpler configuration is better. Positioning is complicated, do not make your life harder by using it too much ⚠️
